+++ example/lib/main.dart
@@ -33,17 +33,11 @@ class _MyAppState extends State<MyApp> {
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Plugin example app'),
-        ),
+        appBar: AppBar(title: const Text('Plugin example app')),
         body: Center(
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              Center(
-                child: Text("scanning: ${scannedValue ?? ""}"),
-              ),
-            ],
+            children: [Center(child: Text("scanning: ${scannedValue ?? ""}"))],
           ),
         ),
       ),